requirements.txt
Django==2.2.2
django-debug-toolbar==2.0
django-environ==0.4.5
mysqlclient==1.4.2.post1
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
